requirements.txt
# ==== LLM / LangChain (latest working versions) ====
langchain>=0.2.10
langchain-core>=0.3.10
langchain-community>=0.2.10
langchain-google-genai>=1.0.3
google-generativeai>=0.7.2

# ==== Environment variable loader ====
python-dotenv>=1.0.1

# ==== MCP server & client ====
mcp[cli]>=1.0.0
httpx>=0.26.0

# ==== Web scraping utilities ====
requests>=2.32.0
beautifulsoup4>=4.12.3

